+++ laravel/routes/web.php
@@ -45,8 +45,10 @@ Route::get('/dashboard', function () {
 // Perfil del Usuario
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
+    Route::patch('/profile/driver-info', [ProfileController::class, 'updateDriverInfo'])->name('profile.driver.update');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+    
 });
 
 // Carga las rutas de login/registro (NO BORRAR)